pyproject.toml
[project]
name = "streamlit_chanu_whale"
version = "0.1.0"
description = "shanu chatbot "
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit", 
    "google-generativeai", 
    "python-dotenv>=1.0.0",
    "dotenv", 
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
    "pre-commit>=3.7.1",
    "ruff>=0.8.0",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff]
line-length = 88
indent-width=4
target-version = "py311"

[tool.ruff.lint]
extend-select = ["I", "U"]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
